feat: :sparkles: Make chatsplit and rename chatgpt
