+++ src/tests/resources/long-code.txt
@@ -1,0 +1,82 @@
+```
+//Include the required libraries and headers
+#include <vulkan/vulkan.h>
+#include <GLFW/glfw3.h>
+#include <vector>
+
+//Define the window dimensions
+const uint32_t WIDTH = 800;
+const uint32_t HEIGHT = 600;
+
+//Defining the instance of the Vulkan program
+VkInstance instance;
+
+int main() {
+    //Initialize GLFW Library
+    glfwInit();
+
+    //Create the GLFW window
+    GLFWwindow* window = glfwCreateWindow(WIDTH, HEIGHT, "Video Editing Program", nullptr, nullptr);
+    if (!window) {
+        glfwTerminate();
+        return -1;
+    }
+
+    //Initialize the Vulkan instance
+    VkApplicationInfo appInfo = {};
+    appInfo.sType = VK_STRUCTURE_TYPE_APPLICATION_INFO;
+    appInfo.pApplicationName = "Vulkan Video Editing Program";
+    appInfo.applicationVersion = VK_MAKE_VERSION(1, 0, 0);
+    appInfo.pEngineName = "No Engine";
+    appInfo.engineVersion = VK_MAKE_VERSION(1, 0, 0);
+    appInfo.apiVersion = VK_API_VERSION_1_0;
+
+    VkInstanceCreateInfo createInfo = {};
+    createInfo.sType = VK_STRUCTURE_TYPE_INSTANCE_CREATE_INFO;
+    createInfo.pApplicationInfo = &appInfo;
+
+    //Enabling required instance extensions and layers
+    std::vector<const char> instanceExtensions;
+    std::vector<const char> instanceLayers;
+
+    //Creating logical
+device
+    VkDevice device;
+    VkPhysicalDevice physicalDevice;
+
+    //Creating swap chain
+    struct SwapChainSupportDetails {
+        VkSurfaceCapabilitiesKHR capabilities;
+        std::vector<VkSurfaceFormatKHR> formats;
+        std::vector<VkPresentModeKHR> presentModes;
+    };
+
+    //Creating render pass
+    VkRenderPass renderPass;
+
+    //Creating graphics pipeline
+    VkPipelineLayout pipelineLayout;
+    VkPipeline graphicsPipeline;
+
+    //Creating command buffers
+    std::vector<VkCommandBuffer> commandBuffers;
+
+    //Creating synchronization objects
+    std::vector<VkSemaphore> imageAvailableSemaphores;
+    std::vector<VkSemaphore> renderFinishedSemaphores;
+    std::vector<VkFence> inFlightFences;
+    std::vector<VkFence> imagesInFlight;
+
+    //Game Loop
+    while (!glfwWindowShouldClose(window)) {
+        glfwPollEvents();
+
+        //Rendering operations go here
+
+    }
+    //Cleanup and cleanup the allocated resources, including Vulkan objects
+    //Cleanup GLFW objects and terminate the application
+
+    return 0;
+}
+```
